# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Alpha Cyber OSINT Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface LoginInput {
  /** @minLength 4 */
  code: string;
}

export type AuthResultRole = typeof AuthResultRole[keyof typeof AuthResultRole];


export const AuthResultRole = {
  admin: 'admin',
  user: 'user',
} as const;

export interface AuthResult {
  success: boolean;
  role: AuthResultRole;
  sessionId: string;
}

export type SessionInfoRole = typeof SessionInfoRole[keyof typeof SessionInfoRole];


export const SessionInfoRole = {
  admin: 'admin',
  user: 'user',
} as const;

export interface SessionInfo {
  sessionId: string;
  role: SessionInfoRole;
  ip?: string;
  userAgent?: string;
  createdAt: string;
  accessCodeLabel?: string;
}

export type AccessCodeRole = typeof AccessCodeRole[keyof typeof AccessCodeRole];


export const AccessCodeRole = {
  admin: 'admin',
  user: 'user',
} as const;

export interface AccessCode {
  id: number;
  code: string;
  /** @nullable */
  label?: string | null;
  role: AccessCodeRole;
  isActive: boolean;
  /** @nullable */
  maxUses?: number | null;
  usedCount: number;
  /** @nullable */
  expiresAt?: string | null;
  createdAt: string;
}

export type AccessCodeInputRole = typeof AccessCodeInputRole[keyof typeof AccessCodeInputRole];


export const AccessCodeInputRole = {
  admin: 'admin',
  user: 'user',
} as const;

export interface AccessCodeInput {
  /** @minLength 4 */
  code: string;
  label?: string;
  role: AccessCodeInputRole;
  /** @nullable */
  maxUses?: number | null;
  /** @nullable */
  expiresAt?: string | null;
}

export interface AccessCodeUpdate {
  label?: string;
  isActive?: boolean;
  /** @nullable */
  maxUses?: number | null;
  /** @nullable */
  expiresAt?: string | null;
}

export interface ActiveSession {
  sessionId: string;
  ip: string;
  userAgent: string;
  /** @nullable */
  country?: string | null;
  /** @nullable */
  accessCodeLabel?: string | null;
  role: string;
  createdAt: string;
  lastSeenAt: string;
}

export interface AuditLog {
  id: number;
  /** @nullable */
  sessionId?: string | null;
  ip: string;
  userAgent: string;
  /** @nullable */
  country?: string | null;
  action: string;
  /** @nullable */
  details?: string | null;
  createdAt: string;
}

export interface AuditLogPage {
  logs: AuditLog[];
  total: number;
  page: number;
  limit: number;
}

export interface PlatformStats {
  totalCodes: number;
  activeCodes: number;
  activeSessions: number;
  totalSearches: number;
  searchesToday: number;
  siteEnabled: boolean;
}

export interface SiteStatusUpdate {
  enabled: boolean;
}

export type SearchInputTargetType = typeof SearchInputTargetType[keyof typeof SearchInputTargetType];


export const SearchInputTargetType = {
  username: 'username',
  email: 'email',
  phone: 'phone',
  name: 'name',
  auto: 'auto',
} as const;

export interface SearchInput {
  /** Username, email, phone number, or full name */
  target: string;
  targetType: SearchInputTargetType;
  /** List of tools to run, or empty for auto-select */
  tools?: string[];
}

export type SearchJobStatus = typeof SearchJobStatus[keyof typeof SearchJobStatus];


export const SearchJobStatus = {
  queued: 'queued',
  running: 'running',
  completed: 'completed',
  failed: 'failed',
} as const;

export interface SearchJob {
  searchId: string;
  target: string;
  targetType: string;
  status: SearchJobStatus;
  tools: string[];
  createdAt: string;
}

export type ToolResultStatus = typeof ToolResultStatus[keyof typeof ToolResultStatus];


export const ToolResultStatus = {
  pending: 'pending',
  running: 'running',
  completed: 'completed',
  failed: 'failed',
  skipped: 'skipped',
} as const;

export interface ResultItem {
  /** @nullable */
  platform?: string | null;
  /** @nullable */
  url?: string | null;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  displayName?: string | null;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  profilePicture?: string | null;
  found: boolean;
  /** @nullable */
  metadata?: string | null;
}

export interface ToolResult {
  tool: string;
  status: ToolResultStatus;
  results: ResultItem[];
  /** @nullable */
  duration?: number | null;
  /** @nullable */
  errorMessage?: string | null;
}

export type SearchDetailsStatus = typeof SearchDetailsStatus[keyof typeof SearchDetailsStatus];


export const SearchDetailsStatus = {
  queued: 'queued',
  running: 'running',
  completed: 'completed',
  failed: 'failed',
} as const;

export interface SearchDetails {
  searchId: string;
  target: string;
  targetType: string;
  status: SearchDetailsStatus;
  tools: string[];
  toolResults: ToolResult[];
  totalFound: number;
  createdAt: string;
  /** @nullable */
  completedAt?: string | null;
}

export type SearchSummaryStatus = typeof SearchSummaryStatus[keyof typeof SearchSummaryStatus];


export const SearchSummaryStatus = {
  queued: 'queued',
  running: 'running',
  completed: 'completed',
  failed: 'failed',
} as const;

export interface SearchSummary {
  searchId: string;
  target: string;
  targetType: string;
  status: SearchSummaryStatus;
  totalFound: number;
  toolCount: number;
  createdAt: string;
}

export interface SearchPage {
  searches: SearchSummary[];
  total: number;
  page: number;
  limit: number;
}

export interface CountByLabel {
  label: string;
  count: number;
}

export interface DailyCount {
  date: string;
  count: number;
}

export interface SearchStats {
  byTargetType: CountByLabel[];
  byStatus: CountByLabel[];
  topTools: CountByLabel[];
  recentActivity: DailyCount[];
}

export type ExportInputFormat = typeof ExportInputFormat[keyof typeof ExportInputFormat];


export const ExportInputFormat = {
  pdf: 'pdf',
  csv: 'csv',
  json: 'json',
} as const;

export interface ExportInput {
  format: ExportInputFormat;
}

export interface ExportResult {
  format: string;
  /** Base64 encoded data or download URL */
  data: string;
  filename: string;
}

export type GetAuditLogsParams = {
page?: number;
limit?: number;
};

export type ListSearchesParams = {
page?: number;
limit?: number;
};