# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Alpha Cyber OSINT Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with access code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "401":
          description: Invalid or expired access code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current session info
      responses:
        "200":
          description: Current session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionInfo"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /admin/codes:
    get:
      operationId: listAccessCodes
      tags: [admin]
      summary: List all access codes
      responses:
        "200":
          description: Access codes list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AccessCode"
    post:
      operationId: createAccessCode
      tags: [admin]
      summary: Create a new access code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccessCodeInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccessCode"
  /admin/codes/{id}:
    patch:
      operationId: updateAccessCode
      tags: [admin]
      summary: Update an access code
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccessCodeUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccessCode"
    delete:
      operationId: deleteAccessCode
      tags: [admin]
      summary: Delete an access code
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /admin/sessions/{sessionId}:
    delete:
      operationId: terminateSession
      tags: [admin]
      summary: Terminate an active session
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Session terminated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/logs:
    get:
      operationId: getAuditLogs
      tags: [admin]
      summary: Get audit logs
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Audit logs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditLogPage"
  /admin/site-status:
    patch:
      operationId: updateSiteStatus
      tags: [admin]
      summary: Enable or disable the site
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SiteStatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
    LoginInput:
      type: object
      properties:
        code:
          type: string
          minLength: 4
      required:
        - code
    AuthResult:
      type: object
      properties:
        success:
          type: boolean
        role:
          type: string
          enum: [admin, user]
        sessionId:
          type: string
      required:
        - success
        - role
        - sessionId
    SessionInfo:
      type: object
      properties:
        sessionId:
          type: string
        role:
          type: string
          enum: [admin, user]
        ip:
          type: string
        userAgent:
          type: string
        createdAt:
          type: string
        accessCodeLabel:
          type: string
      required:
        - sessionId
        - role
        - createdAt
    AccessCode:
      type: object
      properties:
        id:
          type: integer
        code:
          type: string
        label:
          type: ["string", "null"]
        role:
          type: string
          enum: [admin, user]
        isActive:
          type: boolean
        maxUses:
          type: ["integer", "null"]
        usedCount:
          type: integer
        expiresAt:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - code
        - role
        - isActive
        - usedCount
        - createdAt
    AccessCodeInput:
      type: object
      properties:
        code:
          type: string
          minLength: 4
        label:
          type: string
        role:
          type: string
          enum: [admin, user]
        maxUses:
          type: ["integer", "null"]
        expiresAt:
          type: ["string", "null"]
      required:
        - code
        - role
    AccessCodeUpdate:
      type: object
      properties:
        label:
          type: string
        isActive:
          type: boolean
        maxUses:
          type: ["integer", "null"]
        expiresAt:
          type: ["string", "null"]
...
    AuditLog:
      type: object
      properties:
        id:
          type: integer
        sessionId:
          type: ["string", "null"]
        ip:
          type: string
        userAgent:
          type: string
        country:
          type: ["string", "null"]
        action:
          type: string
        details:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - ip
        - userAgent
        - action
        - createdAt
    AuditLogPage:
      type: object
      properties:
        logs:
          type: array
          items:
            $ref: "#/components/schemas/AuditLog"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - logs
        - total
        - page
        - limit
...
    SiteStatusUpdate:
      type: object
      properties:
        enabled:
          type: boolean
      required:
        - enabled
...