# lib/api-zod/src/generated/types/accessCode.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Alpha Cyber OSINT Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { AccessCodeRole } from './accessCodeRole';

export interface AccessCode {
  id: number;
  code: string;
  /** @nullable */
  label?: string | null;
  role: AccessCodeRole;
  isActive: boolean;
  /** @nullable */
  maxUses?: number | null;
  usedCount: number;
  /** @nullable */
  expiresAt?: string | null;
  createdAt: string;
}