# lib/api-zod/src/generated/types/accessCodeUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Alpha Cyber OSINT Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface AccessCodeUpdate {
  label?: string;
  isActive?: boolean;
  /** @nullable */
  maxUses?: number | null;
  /** @nullable */
  expiresAt?: string | null;
}