# lib/api-zod/src/generated/types/auditLog.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Alpha Cyber OSINT Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface AuditLog {
  id: number;
  /** @nullable */
  sessionId?: string | null;
  ip: string;
  userAgent: string;
  /** @nullable */
  country?: string | null;
  action: string;
  /** @nullable */
  details?: string | null;
  createdAt: string;
}