# lib/api-zod/src/generated/types/sessionInfo.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Alpha Cyber OSINT Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { SessionInfoRole } from './sessionInfoRole';

export interface SessionInfo {
  sessionId: string;
  role: SessionInfoRole;
  ip?: string;
  userAgent?: string;
  createdAt: string;
  accessCodeLabel?: string;
}